flask_chat/requirements.txt
# Flask: framework web leve usado para criar a aplicação
Flask==2.3.3

# Flask-SocketIO: permite comunicação em tempo real via WebSockets
Flask-SocketIO==5.3.4

# python-dotenv: carrega variáveis de ambiente do arquivo .env (como API keys)
python-dotenv==1.0.1

# eventlet: servidor assíncrono necessário para suportar WebSocket com Flask-SocketIO
eventlet==0.35.1

# --- BIBLIOTECAS ATUALIZADAS E CORRIGIDAS ---
google-generativeai==0.5.4
langchain==0.2.1
langchain-google-genai==1.0.5
Pillow==10.3.0
# --- FIM DAS ATUALIZAÇÕES ---

gTTS
gunicorn